More than one keyring is needed to verify.
gpg has a very annoying attitude:

If more then one key is used in the signature,
all signatures must be perfectly verified,
else it bails out with a fatal error (code 2).

So if you have one good and one bad key,
the result is fatal, hence the .sign file
fails signature check.

I'd rather would like to be able to specify,
that one good signature is enough.  But AFAICS
this is not implemented in gpg.  As always.

And no, I do not grep 'Good signature' or similar
in the output of gpg, because then somebody might
create a key named 'Good signature' (or whatever
I grep for) and add it as a BAD key but grep then
is successful.  Nope.  No way.  Either the crypto
tool is usable, or (like gpg) it is not.

Hence the requirement now is that all keys pass.

But .. the keys of Ubuntu 14.04.1 are in different
keyrings now.  Luckily, we can give multiple
keyrings to gpg.  I hope this isn't just pure luck
that this feature exists so it will prevail.

So what I now do is just give gpg all keyrings
I can find, alltogether.  This then solves
our multiple signature problem.

For now.  Let's see what future brings ..